+++ extensions/response.py
@@ -2,10 +2,6 @@ from typing import Any
 
 from extensions import logger
 
-from conf.const import StatusCode
-
-
-
 
 def resp_success(message: str = '', print_msg: str = '', data: Any = None):
     if print_msg: