+++ redis_ext/sms.py
@@ -6,7 +6,3 @@ class SMSCodeRedis(BaseRedisClient):
 
     DB = 1
     PREFIX_KEY = 'sms_code'
-
-    def __init__(self, key) -> None:
-        super().__init__(key)
-        self.key = f'{self.PREFIX_KEY}:{key}'