+++ redis_ext/token.py
@@ -6,28 +6,3 @@ class TokenRedis(BaseRedisClient):
 
     DB = 1
     PREFIX_KEY = 'token'
-
-    def __init__(self, phone: str, extend_model: str, extend_user_id: int) -> None:
-        super().__init__(f'{phone}:{extend_model}:{extend_user_id}')
-        self.key = f'{self.PREFIX_KEY}:{phone}:{extend_model}:{extend_user_id}'
-
-    def set_login_info(self, mapping: dict):
-        """set token
-
-        mapping: key: value
-            {
-                "login_at": float,
-                "token_expired": float,
-                "extend_user_id": int
-            }
-        """
-
-        return self.client.hmset(self.key, mapping=mapping)
-
-    def get_login_info(self, keys: list):
-        """get token
-
-        keys: [key]
-        """
-
-        return self.client.hmget(self.key, keys=keys)